Fixed some bugs with the scraping
Fixed a bug when scraping https://www.youtube.com/watch?v=sp1rkgx_aik.
There was an issue with a timeout on trying to get the next comment.
Sometimes the listed number of comments doesn't match the actual number,
so you have to handle an exception of grabbing the next comment (i.e.
the selector you use does not match any comment on the page, since a
comment is deleted) since you never know which comment is actually the
last comment (compared to the listed number of comments on a Youtube
video).