Fixed issue where changes in setup bash script were applied globally
There was an issue with the bash script where the changes with pip and
python were applied globally. So I used "/usr/bin/env python -m" as a
prefix when using pip. This way, the version of python and pip used is
the one in the virtual environment, and the changes are only in the
virtual environment (and not global).