Made changes to the bash script to be more flexible
Added help messaging for the bash script, as well as the ability to
specify arguments. Now the script can specify a help message with the -h
argument (or --help or -help), and it can specify a name for the virtual
environment directory if you want to use a different name than the
default. Run "./setup --help" for help messaging.

Started adding the argument parser in the main.py file to parse
arguments using the argparse module.

Disclaimer: I used the below stack overflow answer for help on how to
use getopt.
https://stackoverflow.com/questions/16483119/an-example-of-how-to-use-getopts-in-bash