Added a flag to specify caching behaviour
Added a flag to specify caching behaviour called --nocache (can also be
passed in as -n or -nocache). If specified, caching is disabled. When
running pip install, the flag --no-cache-dir is specified. When running
pip uninstall to remove dependencies that differ from the
requirements.txt file, the command "pip cache purge" is also run to
purge the local cache. The new changes support this flag.