Added a script to download the relevant version of chromedriver
Added a script to automatically download the latest stable version of
chromedriver that is specific to the operating system and hardware of
the machine the script is running on. For example, if you are running
this on a 64 bit windows machine, then the script will download the
latest stable version of chromedriver for 64-bit windows machines and
add it to your ~/bin/ directory (because in Git Bash, ~/bin/ is part of
your path).

Next steps are to add a sanity test to ensure that the binary is
executable on the machine the script is running on, as well as to double
check that ~/bin/ is part of the PATH variable for the operating system.
If it is not, the script can add ~/bin/ to the path to ensure that the
chromedriver is in your path. It may also be worth looking into a custom
name for a directory that is added to the home directory, and then
adding the chromedriver there, and adding this custom-named directory as
part of the path. This could ensure that it doesn't clash with the
user's ~/bin/ directory which could already be in their $PATH.