Removed chromedriver download process from Bash script
Removed the chromedriver download process from the Bash script, instead
offloading it onto a separate python script called "get_webdriver.py"
that will run and install the relevant version of chromedriver based on
a given platform name.

The platform name is determined in this script using the uname command
to get the operating system name and the machine architecture name. The
script will then determine if you have a valid platform that is supported
by chrome webdriver.

The script will first check to see if you have chromedriver installed in
your ~/bin/ directory, and if it is of the latest version. If it is,
then don't do anything. Otherwise, the script will find out the relevant
platform name and accordingly download the relevant version of chrome
webdriver. From there, the script will unzip only the chromedriver
binary.

To do:
* add the newly installed chromedriver binary to the ~/bin/ directory
* add safeguards to ensure that the ~/bin/ directory is part of the
  path
* connect this script to setup.sh. Add documentation for this new
  functionality
* Consider adding a custom directory to your path that contains
  the downloaded chromedriver binary, and add functionality to remove
  this custom directory as well after you have used the script (by
  adding a flag in setup.sh to add the binary to a custom directory that
  will be added to path, and then adding a flag to remove that custom
  directory)

Next commit:
* python script that downloads the chromedriver binary