Offset the downloading of chromedriver binary to a Python script
Offset the downloading of the chromedriver binary to this Python script.
The script will take in one argument for the platform, and then download
the latest stable chromedriver version. If things go wrong, it will
print error messages based on the issue and exit with a non-zero exit
code. If everything goes well, it will print out the name of the
downloaded zip file and then exit with an exit code of 0. The downloaded
zip file name comes from the endpoint where we are downloading the zip
file. Everything after the last slash corresponds to the name of the
local zip file.