Fixed the download_chromedriver.sh and get_webdriver.py scripts
The script was not downloading the executable, so after lots of
debugging, it was found that the python script was not downloading the
chromedriver zip file in the right directory, and there were issues with
the bash commands I used in the script. So the python script now
downloads the zip file directly in the ~/bin/ directory. This way the
zip file is immediately accessible by the script. There were issues in
the bash code with variable names being out of date, regular expressions
being wrong, and others. So now the script handles the bin/ directory
not being there, the bin/ directory being there but ther being no
chromedriver executable, or the chromedriver executable being there but
the version is not up to date (or the file does not print out the
version and is just named the same as the chromedriver executable should
be named).