Fixed 2 bugs with the scraping of comments
In the YouTube video page, there is a list of the total number of
comments. For videos with large numbers of comments (thousands), the
number has a comma every 3 digit places. So I added code to account for
this and remove the commas from the text that resembles the number of
comments. I also changed the conditional after which you stop scraping
comments to stop if the number of scraped comments is greater than or
equal to the limit.