Adding support for specified time limits
I'm now adding support for time limits you can specify to have the
script run for a certain time limit. The number of hours, minutes and
seconds is combined to get the total second count that you want the
script to run. The script won't run for that exact number of seconds. It
will finish parsing the current comment (or current reply in a comment
thread) and then stop scraping the rest of the comments on the video.

I am also adding verification to ensure that the parameters passed in
are valid inputs (i.e. comment limits are >= 0, time limits are > 0 if
specified, etc.). There is also an issue with the bash setup script on
the part where you download the latest version of chromedriver. I will
be debugging this next.