Added more parameter validation, and time limit functionnality
Added more parameter validation to validate some of the other
parameters, and added the time limit functionality. Added functionality
to match passed in links with youtube video links based on a regular
expression pattern. Also, the passed in hours, minutes and seconds
are now 0 by default. If you pass in only seconds as a non-zero value,
it has to be 30 secods or more. Otherwise a message is printed to standard
error, and the exit code is 1. If you don't pass in any time limit
parameters (i.e. hours, minutes, seconds), or pass them all in as 0, the
script assumes there is no time limit on how long the script can run.