Fixed an issue where downloading chromedriver didn't work
There was an issue where a variable was referenced before it was
assigned a value, causing the bash script to run a python script with
the wrong path. This caused chromedriver to not get downloaded, and
this wasn't caught because there was no if conditional to catch one edge
case that would be an error. I added an if case to catch this error case
where an empty string is returned by the python script (meaning
something went wrong in invoking it).