Fixed linux-specific issues with the bash script
I fixed some of the issues that related to the bash script when run on
linux. One was that to get the kernel name, I ran the uname command with
the wrong flag. Another was that when virtual environment directories
are created in linux, the path to the activate script is different. So
I found a cross-platform way to run the activate script. This required
checking that the script is there and accounting for the case where it
isn't, and checking that we can source it successfully (and accounting
for the case where this isn't true as well). There are still some bugs
when running this script in linux, but it will work.