Plugged in the time limit functionality into the main script
Fixed the time limit functionality so that the script will now work with
time limits. Updated the checking of passed in parameters since the
default value for hours, minutes and seconds is now 0. So therefore
there is no point in checking that the parameters are not None. When
plugging in the new functionality, there was a bug where 30 seconds
accidentally registered as 1800 seconds. This was because parameters
were accidentally switched in the function declaration, causing the
wrong time limit to be accumulated.