Design pattern change to use the factory pattern
Design pattern change to use the factory pattern to invoke the iterator,
allowing for the codebase to be maintainable as this project scales up.
The steps for scraping YouTube shorts is different, so a different
iterator would be necessary for YouTube shorts. All iterators will take
the url as the first argument, and will have a static method that
returns a regex string pattern. This regex pattern is used in the
factory class to match the given URL passed in, so that you can then
dynamically retrieve the correct iterator based on the passed in URL,
and this solution can scale well for multiple different social media
sites (i.e. Instagram, TikTok, etc.).