Added command line flags to specify logging and a logfile
modifed the log messages in the logging decorator, put the decorator on
a different function, and added command line flags to specify logging
and an output logfile. I also used the vars() builtin function to turn
the arguments into a dictionary of keyword arguments. These keyword
arguments are passed into the instantation of the IteratorFactory class.