Fixed an issue with comment replies not being found
There was a bug with comment replies not being found. The issue was that
I was trying to find and click on the wrong selector, causing the except
block in self.iterate_child to exit prematurely. I forgot that regardless
if comment replies are found, the button to hide the replies has a different
CSS selector than the button to show the replies. Having the finally block
start sooner is what allowed for us to handle the edge case where the button
to show comment replies exists, but the replies themselves do not exist.