Added code for iterating over comments for youtube shorts
Added code for iterating over comments for youtube shorts, but the code
has bugs in it that have to get fixed. After a certain while, the next
comment is not found for some reason. This may be due to the amount that
is being scrolled. Hence, the implementation for iterating over YouTube
shorts is not yet ready, and still needs work.

Removed commented out code in bash script.