Made changes to the scrolling function, other small changes
Changed the scrolling function to not use a default margin, and to
scroll by the element height plus the total margins. I added this
because there was a bug related to scrolling. The div was not being
scrolled enough due to an earlier bug with margins. I was taking a
margin of 20 off each time that I scrolled, so that each time the
scrollbar scrolled, it was falling behind by 20 pixels.

To illustrate, if the first element has a height of 100 pixels, then
I scroll down by 80. If the second element has a height of 60 pixels,
then I scroll down by 40 pixels. However, when I get to examining the
third element but I haven't scrolled for a third time, the third
element is offset from the top of the container div by 40 pixels. By the
time I get to the fourth element, the element is offset from the top of
the container div by 60 pixels, and so on. This would compound till I
would get to an element that is out of the screen and doesn't load
because of the componding offset of pixels from the top of the container
div. So to get it to work, I kept track of the element height, along
with the margin-top and margin-bottom css properties. I still have to
account for the border width, and padding as well to ensure that all
aspects of the css box model are accounted for.

I added an optional parameter in the get_selector method to specify wait
time if you want a different wait time from the default (currently 10
seconds). Other changes were made to the logging of exceptions. I added
some exception safety and some logging of errors in the __next__ method,
and there were some other minor code changes here and there (commenting
out of breakpoint() calls, commenting out of dummy conditionals used for
testing, etc.).