Accounted for the border width and padding width in the scroll height
When each comment thread is scrolled down, the height, along with the
vertical padding, vertical margins, and border widths all have to be
accounted for. I added code to account for all of them. I also added
some docstrings for some functions, added an optional wait_time
parameter in the element_exists method, and refactored some code. I
added some logging in the element_exists method as well. I tried to have
the iterator log exceptions only when the element exists, as otherwise
everytime the iterator passes the last comment there will be an
exception logged in the logfile. I did some refactoring of code as well.