Implemented a method of writing JSON to disk during iteration
For YouTube videos with large numbers of comments, creating a dictionary
with the list of all comments is not feasible since your computer will
run out of memory. As an example, some YouTube videos have hundreds of
thousands of comments, and sometimes millions. If you try to add all of
the comments into a long list of dictionaries, then you have to use use
json.dumps to convert it to a JSON string. Then you have to try to write
that large string to your json file. For videos with lots of comments you
may not have enough RAM memory in your machine. So I made an iterator
that inherits from the list class, and passed into the json.dump method.
This allows dictionaries to be converted into JSON objects and get
written to the file as the dictionaries are returned by the iterator.
This ensures you never have to read in all of the dictionaries of comments
at once.

To clarify, I did not come up with this myself. I copied the code from a
stack overflow answer (below).
https://stackoverflow.com/questions/52137608/writing-a-large-json-array-to-file