Attempt to make sure the buffer solution works
Tested the same video, once with the buffer flag on and once with it
off. I found no difference in seconds passed when the driver closed, and
when the program finished running (when counting in my head). So I tried
to add a yield statement in the __next__ method in the hopes that making
the IteratorAsList class a generator would cause the dictionaries to be
written to the JSON file more frequently during iteration (instead of at
the end). I tested it once more (once with the buffer flag and once
without the buffer flag for the same video), and noticed a slight
difference in counted seconds from when the driver closed and when the
program finished execution.