Fixed comment scrolling issue
I fixed a comment scrolling issue where the amount that was being
scrolled down did not cover enough ground by a small amount. This
marginal amount that the scrollbar would not scroll down by would
compound to the point where you would have the comments that are
currently being parsed where they may eventually fall out of the screen.
This would cause your web scraper to stop prematurely, and it would
cause the tool to fail for videos with large numbers of comments.

I used a JavaScript method that exists on DOM element objects called
"scrollIntoView" that would scroll the element in question to the
top of the containing div. This would allow for comments to always be
scrolled to the top even when there are thousands of comments.

I refactored the code according to this new change. I updated the
docstring for the method in question, and removed code that wasn't
needed.

There is now an error where the last comment is not being collected for
some reason. Instead, the action of pressing the down arrow key seems to
be mimicked since the last comment is not accounted for and the next
video is shown (which happens by pressing the down arrow key). I will
have to investigate why the last comment is not being read in.