Slightly fixed an issue where down arrow key is accidentally pressed
There was an issue with youtube shorts where the scrollIntoView method
accidentally had the same effect as the down arrow. This would happen
once you scrolled the comments fully down to the bottom of a youtube
shorts video. It would also happen randomly at other times.

The effect was that the last one or 2 comment threads might not be
scraped and included in the resulting JSON file. You could also be
scraping a large number of comments on a YouTube shorts video (a few
thousand for example). and then you suddenly go to the next video.

Going to the next YouTube shorts video when you are currently watching a
short video is done by pressing the down arrow key. For some reason, the
scrollIntoView method seems to have the same behaviour as someone
pressing the down arrow key. I made a duct-tape fix by checking the
amount left to scroll, and ensuring that the amount left to scroll is
not very small (i.e. less than 10 or 20 pixels). When the amount left to
scroll is less than the threshold, then you try to hold the scrollbar of
the containing element that contains all comment threads and you try to
drag the scrollbar down. This seems to work for some YouTube short
videos, but has not been thoroughly tested. There is an edge case where
this also doesn't work (one specific video, linked below):
https://www.youtube.com/shorts/R_j9cSiMolE .