Fixed the bug related to accidentally scrolling down to next video
I took a different approach where I tried setting the style of the
scrollbar using javascript. Then depending on where the current element
is in the visible container (i.e. the bottom half or the bottom third),
you either use the scrollIntoView method or you add the scrollHeight of
the comment box container to the scrollTop property of the current
element. Originally the plan was to find the exact coordinates of the
down arrow button at the bottom part of the scrollbar, and then click
and hold the button for a second. However, when I tried the old code
with the newly styled scrollbar I found that my old code did not
accidentally scroll down to the next YouTube shorts video (i.e. the part
of adding the container's scrollHeight value to the current element's
scrollTop value). So I made changing the scrollbar style be part of the
setup. Instead, if the comment in question is in the bottom half of the
visible container, then the javascript code to scroll the element will
involve adding to the current element's scrollTop value.