Added test to test the limit of 500 comments
Added a test to test a limit of 500 comments. I also refactored the code
to have a decorator function to check if the video is there, and then
skip the tests if the video is not there. This would be applied as a
decorator to your functions and can save you some lines of code.