Fixed a bug where time limits were not working
I fixed a bug where time limits were not working because the
self.seconds, self.minutes and self.hours properties were being assigned
0 in the __init__ function. This was causing the time limit to never be
set, which was causing any passed in parameters for time limits to be
ignored. Now it works when specifying seconds, but has not been tested
when specifying minutes or hours.

I added a test to test the 30 second time limit.