Added default logfile test
Added a test to test that when enabled_logging is specified to True, the
logfile is created. I ensured that when running the setup, any files
with the name "debug.log" are deleted from the current directory. I
wrote the code such that it should work in both Windows and Linux.