Added tests for iterators for regular YouTube video comments
Added tests for iterators for regular YouTube video comments, and made
changes to the CommentIterator codebase to ensure that these tests pass.
I tried to find a way to make the test class for the regular comments
inherit from the test class for YouTube shorts, but I couldn't get the
tests for regular YouTube videos to run standalone. The tests for
YouTube shorts would run after the tests for regular YouTube videos
would run.