Refactored the test classes and added a bash script as a test harness
After experimenting I found out that I did not need to copy all the
methods over into the new test class for regular YouTube videos. The
original reason why both the tests for regular YouTube videos and for
YouTube shorts were running was because I imported the class from
the module where it was defined. It seemed to cause the tests to run
for some reason. When I didn't import the class, only the tests for
regular YouTube videos ran. From here I realized that I could refactor
the test classes for regular YouTube videos to inherit from the test
classes for YouTube shorts. All that needed changing between the classes
was the instantiation of some attributes in the __init__ constructor for
both the superclasses and subclasses.

More so, when I ran the tests in the form of
"python -m unittest -v package_name.module_name.ClassName" only the
specific test class would run by itself. This prompted me to create a
bash script called run_tests.sh that served as a test harness to run the
various tests standalone.