Added functionality to read settings from JSON file
Earlier this script could scrape videos by specifying arguments through
the command line. Now you can specify a configuration file with the
--configfile command line argument. This argument specifies a JSON file
that will contain an array of JSON objects representing each video to
scrape. To invoke the script, you now have to specify one of the two
arguments in the command line:
1) the --url argument.
2) the --configfile argument.
See the below example for how the JSON in the config file should be structured.
{
	"videos": [
		{
			"url": "https://www.youtube.com/shorts/7uctTsKeLdM",
			"output": "comments.json",
			"enabled_logging": true,
			"logfile": "debug.log",
			"limit": 500000,
			"minutes": 30
		},
		{
			"url": "https://www.youtube.com/watch?v=1qG_FQbTZ3U",
			"output": "basketball_video.json",
			"enabled_logging": true,
			"logfile": "basketball_video.log",
			"limit": 500000,
			"buffer": true,
			"minutes": 30
		}
	]
}